requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==3.1.4
gunicorn==20.0.4
djangorestframework
django-appconf==1.0.3
django-crontab==0.7.1
django-filter==2.4.0
django-audit-log==0.7.0
django-contrib-comments==1.9.1
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-debug-toolbar==2.1
django-formtools==2.1
django-staticfiles==1.2.1
django-treebeard==4.3
psycopg2==2.9.1
whitenoise
pandas==1.3.4